test: RecommendedArtists
